+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+
+- `.` without following identifier wrongly recognized as valid local label
+          
+
 ## 0.1.0
 
 ### Added